requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic
python-multipart==0.0.9
httpx==0.27.2
python-docx==1.1.2
python-dotenv>=1.0.0
google-generativeai>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
